+++ src/NumSharp.Core/Backends/Iterators/NDIteratorCasts/NDIterator.Cast.UInt16.cs
@@ -76,7 +76,7 @@ namespace NumSharp
                         {
                             var hasNext = new Reference<bool>(true);
                             var iterator = new ValueCoordinatesIncrementor(ref shape, delegate(ref ValueCoordinatesIncrementor _) { hasNext.Value = false; });
-                            Func<int[], int> getOffset = shape.GetOffset;
+                            Func<long[], long> getOffset = shape.GetOffset;
                             var index = iterator.Index;
 
                             MoveNext = () =>
@@ -197,7 +197,7 @@ namespace NumSharp
                         {
                             var iterator = new ValueCoordinatesIncrementor(ref shape, delegate(ref ValueCoordinatesIncrementor incr) { incr.Reset(); });
                             var index = iterator.Index;
-                            Func<int[], int> getOffset = shape.GetOffset;
+                            Func<long[], long> getOffset = shape.GetOffset;
                             MoveNext = () =>
                             {
                                 var ret = convert(*((UInt16*)localBlock.Address + getOffset(index)));
